separate comb_init and comb_clear
